Move packagejson resolver implementation to new package
